requirements.txt

streamlit
ollama
Pillow
pdf2image
docxtpl
